arm64: dts: SoMC: Tama: Update wled scale current limit for Apollo
The WLED driver found in the current kernel version only supports
setting power values that are predefined in the driver itself in the
wled_fs_current_values array. so let's use 20000 uA for Apollo as well
as Akari. Also move the configuration to Tama common dts and disable the
WLED driver for Akatsuki device because it has an OLED panel.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>